requirements.txt
anthropic
langchain
langchain-anthropic
langchain-community
langchain-chroma
chromadb
pypdf
streamlit
python-dotenv
tiktoken